% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{calcCSS}
\alias{calcCSS}
\title{Identifying cell-type specific transcription factor regulons}
\usage{
calcCSS(
  mat,
  pos_cells,
  use_cluster = NULL,
  neg_cells = NULL,
  scale = FALSE,
  nperm = 1000,
  ncores = 1,
  verbose = interactive()
)
}
\arguments{
\item{mat}{A numeric \code{matrix} contains transcription factor regulons activate score.}

\item{pos_cells}{Cells from the cluster(s) to calculate the cell-type specific
score. It must be included in column names of mat.}

\item{use_cluster}{The cluster used for cell-type specific calculation.
Default: NULL.}

\item{neg_cells}{Cells from the control cluster(s) to calculate the cell-type
specific score. It must be included in column names of mat. Default: NULL,
all cells expect for pos_cells.}

\item{scale}{Whether to scale values for calculation. If the activate scores
have negative values, it must be scaled. Default: FALSE.}

\item{nperm}{Number of permutations used for calculating p-values against CSS.
Default:1000.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}

\item{verbose}{Whether to output message? Default: TRUE.}
}
\value{
A \code{data.frame} with differential CSS for each TFRs.
}
\description{
Identifying cell-type specific transcription factor regulons
}
