% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonPathwaySimilarity.R
\name{calcuModuleSimilar}
\alias{calcuModuleSimilar}
\title{Estimating the jaccard similarity between two gene lists}
\usage{
calcuModuleSimilar(
  tfr_list,
  pathway_list,
  return_items = c("all", "similarity"),
  method = c("asymptotic", "bootstrap", "mca", "exact"),
  nperm = 1000,
  ncores = 1
)
}
\arguments{
\item{tfr_list}{A list of transcription factor regulons. Names of the list were TF genes,
while the value of each list is its target genes.}

\item{pathway_list}{A list of pathway gene sets. Names of the list were pathway names,
while the value of each pathway is genes.}

\item{return_items}{The return value of similarity values. Default: "all".}

\item{method}{Method used compute p-values in \pkg{jaccard} package. The `asymptotic`
is the fastest, while the `bootstrap` is better but a little bit slow. The `exact`
solution is slow for a moderately large vector. Default: `asymptotic`.}

\item{nperm}{Number of . Used only for `bootstrap` method.}

\item{ncores}{Number of cores used for parallel calculation.}
}
\value{
a list of similarity test \code{data.frame} or a \code{matrix} of
normalized similarity scores.
}
\description{
This method estimate the similarities between two gene lists by using
Jaccard/Tanimoto similarity coefficients.
}
