% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonPathwaySimilarity.R
\name{collapse_module}
\alias{collapse_module}
\title{Merging Enriched pathways based on term similarity}
\usage{
collapse_module(
  enrich_tab,
  key_ids = "pathway",
  module_name = "overlap_targets",
  group_by = c("padj", "similarity", "pval"),
  fdr = 1L,
  similarity_method = c("Jaccard", "Ochiai"),
  min_cluster = 5,
  max_cluster = "auto",
  verbose = interactive(),
  ...
)
}
\arguments{
\item{enrich_tab}{The enrichment result table. It must be a \code{\link{data.table}}
or \code{\link{tibble}}.}

\item{key_ids}{The key column to merge, usually enriched terms, it must be unique.}

\item{module_name}{The enriched genes of each term in key_ids column. It must 
be a list. So the input 'enrich_tab' could only be a \code{\link{data.table}}
or \code{\link{tibble}}.}

\item{group_by}{The value used to represent the enriched term in each module.
Default: padj.}

\item{fdr}{The cutoff used filtering significantly enriched terms. Default: 1, 
no filtering.}

\item{similarity_method}{The method used to calculate the similarity between
enriched terms.}

\item{min_cluster}{The minimum number passed to cluster terms, terms small 
than this value will not be clustered. Default: 5.}

\item{max_cluster}{The maximum number of cluster after clustering. Default: 
"auto", \code{sqrt(nrow(enrich_tab))}.}

\item{verbose}{Whether to output information while running?}

\item{...}{Other parameters passed to \code{\link{eclust()}}}
}
\description{
The \code{collapse_module} function is designed to clustering redundant terms 
into several modules based on the Jaccard similarity of enriched gene. The
term with minimum adjusted p-value in each module used to represent this 
module.
}
