% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfToList}
\alias{dfToList}
\title{Transform data.frame to gene list}
\usage{
dfToList(df, tf_col = "tf", target_col = "target")
}
\arguments{
\item{df}{A data.frame with TF and it's target information.}

\item{tf_col}{The column names of TFs.}

\item{target_col}{The column names of targets.}
}
\value{
A list of Transcription factor regulons. The names of list is TFs
}
\description{
Transform data.frame to gene list
}
\examples{
data(dorothea_regulons)
mouse_regulons <- dfToList(df=dorothea_regulons$mm_regulons,
tf_col="tf", target_col="target")

}
