% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{do_RadvizFeaturePlot}
\alias{do_RadvizFeaturePlot}
\title{do_RadvizFeaturePlot}
\usage{
do_RadvizFeaturePlot(
  rv,
  use_feature,
  outline_circle = TRUE,
  plot_type = c("point", "hexagonal"),
  colors = scale_color_gradient(low = "grey80", high = "dodgerblue4"),
  text_size = 8
)
}
\arguments{
\item{rv}{A \code{\link{getRadviz}} object generate from \code{\link{do.radviz}}}

\item{use_feature}{Features used to split the screens.}

\item{outline_circle}{Whether to add circle in figure?}

\item{plot_type}{Figure types to plot.}

\item{colors}{Colors for gradient.}

\item{text_size}{Text size.}
}
\value{
a ggplot object
}
\description{
do_RadvizFeaturePlot
}
