% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulons.R
\name{doMotifRankGSEA}
\alias{doMotifRankGSEA}
\title{Performing GSEA with motif ranking data}
\usage{
doMotifRankGSEA(
  motif_rank,
  gene_list,
  motif_tf_anno,
  anno_type = c("directAnnotation", "inferred_Orthology", "inferred_MotifSimil"),
  p_value = 0.05,
  revValue = TRUE,
  ncores = 1,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{motif_rank}{A \code{\link{data.frame}} containing motif ranking matrix,
in which rows corresponding to motifs and columns corresponding to genes. Note, The
first rows were motif ids}

\item{gene_list}{A list of pre-defined modules. Names of each mdule should be TFs.}

\item{motif_tf_anno}{A \code{\link{data.table}} with motif binding TF
information download from \pkg{\link{RcisTarget}}.}

\item{anno_type}{The annotation type of motifs with TFs, should be some of:
"directAnnotation", "inferred_Orthology", "inferred_MotifSimil". Default:
all above 3 types.}

\item{p_value}{set the cutoff of significant enriched motifs. Default: 0.05.}

\item{revValue}{Should reverse the values before perform GSEA?}

\item{ncores}{Number of cores used for parallel calculation. Default: 1.}

\item{verbose}{No messages should print while running.}

\item{...}{Others parameters passed to \code{\link{fgsea}}.}
}
\value{
A \code{\link{data.table}} with motif GSEA result.
}
\description{
Performing GSEA with motif ranking data
}
