% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerGRNs.R
\name{filterNets}
\alias{filterNets}
\title{Filtering the weighted regulatory network}
\usage{
filterNets(df, minSize = 100L, maxSize = 5000L, maxInter = 1L, ncores = 1L)
}
\arguments{
\item{df}{A data.frame with only one regulator}

\item{minSize}{Minimum number of targets to keep. Default: 100.}

\item{maxSize}{Maximum number of targets to filter. Default: 5000.}

\item{maxInter}{Maximum number of iteration. At least 1. Default: 5.}

\item{ncores}{Number of cores to use. Default: 1.}
}
\value{
A weighted \code{data.frame} with regulator and target information.
}
\description{
Filtering the regulatory network by evaluating the threshold of each regulator
based on the distribution of weight.
}
\examples{
net <- readRDS(system.file("./inst/extdata/mm_pantissue_network.rds",package = "scATFR"))
net <- head(net,100000)
net <- filterNets(df=net)

}
