% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRegulons.R
\name{filterRegulonByGSEA}
\alias{filterRegulonByGSEA}
\title{Filtering Transcription factor regulons using GSEA}
\usage{
filterRegulonByGSEA(
  regulons,
  grn_tab,
  score_type = c("std", "pos", "neg"),
  normalize = FALSE,
  withWeight = TRUE,
  minSize = 5L,
  maxSize = 1000L,
  filter_pval = 1L,
  ncores = 1
)
}
\arguments{
\item{regulons}{Gene list of regulons, the names of list must be transcription
factor, the values of each list should be gene symbols.}

\item{grn_tab}{A matrix contains weighted gene regulatory network.}

\item{score_type}{This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")}

\item{normalize}{Whether to normalize the ranking score?}

\item{withWeight}{Whether to keep weight in output regulons?}

\item{minSize}{The minimum number of genes in one TFR.}

\item{maxSize}{The maximum number of genes in one TFR.}

\item{filter_pval}{Set the cutoff of significant enriched TFRs. Default:1(no filtering).}

\item{ncores}{number of cores used.}
}
\value{
A gene list with regulon information.
}
\description{
Filtering Transcription factor regulons using GSEA
}
