% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRegulons.R
\name{filterRegulons,SingleCellExperiment-method}
\alias{filterRegulons,SingleCellExperiment-method}
\title{Filtering Transcription factor regulons using Gene regulon networks}
\usage{
\S4method{filterRegulons}{SingleCellExperiment}(
  x,
  gene_list = NULL,
  grn_mat = NULL,
  use_regulon = 1L,
  use_grn = 1L,
  score_type = c("std", "pos", "neg"),
  normalize = FALSE,
  minSize = 5L,
  maxSize = 500L,
  filter_pval = 1L,
  ncores = 1,
  ...
)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} object.}

\item{gene_list}{The regulon list used for filtering. default: NULL(use GRNs in your object).}

\item{grn_mat}{The gene regulatory network used for filtering.}

\item{use_regulon}{Regulons used for filtering.}

\item{use_grn}{GRNs used for regulon filtering.}

\item{score_type}{This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")}

\item{normalize}{Whether to normalize the ranking score?}

\item{minSize}{The minimum number of genes in one TFR.}

\item{maxSize}{The maximum number of genes in one TFR.}

\item{filter_pval}{Set the cutoff of significant enriched TFRs. Default:1(no filtering).}

\item{ncores}{number of cores used.}

\item{withWeight}{Whether to keep weight in output regulons?}
}
\value{
An atfr object.
}
\description{
Filtering Transcription factor regulons using Gene regulon networks
}
\examples{

}
