% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{findCellTypeSignature,matrix-method}
\alias{findCellTypeSignature,matrix-method}
\alias{findCellTypeSignature,SingleCellExperiment-method}
\title{Identifying cell-type specific transcription factor regulons for indicated cluster}
\usage{
\S4method{findCellTypeSignature}{matrix}(
  x,
  pos_clust,
  col_data,
  clust_name,
  neg_clust = NULL,
  scale = FALSE,
  nperm = 1000,
  ncores = 1,
  verbose = interactive()
)

\S4method{findCellTypeSignature}{SingleCellExperiment}(
  x,
  pos_clust,
  clust_name,
  neg_clust = NULL,
  use_altExp = TRUE,
  use_assay = 1L,
  scale = FALSE,
  expression_per = 0.1,
  nperm = 1000,
  ncores = 1,
  verbose = interactive()
)
}
\arguments{
\item{x}{An object with gene expression data used for regulon activity evaluation.
Could be \code{matrix}, \code{SingleCellExperiment}.}

\item{pos_clust}{Indicated cluster used to calculate the cell-type specific score (CSS).}

\item{col_data}{A \code{data.frame} with cell cluster information. Each row
represent a cell and each column represent a cell metadata. The row names of
col_data must be identical with column names of x.}

\item{clust_name}{One column name in colData(x)(for \code{SingleCellExperiment})
or col_data (for \code{matrix}) specific the clusters used for cell-type
specific score calculation.}

\item{neg_clust}{Indicated cluster used as ccntrol set to calculate the
cell-type specific score (CSS) for. Default: NULL, all others cell clusters
except for pos_clust.}

\item{scale}{Whether to scale values for calculation. If the activate scores
have negative values, it must be scaled. Default: FALSE.}

\item{nperm}{Number of permutations used for calculating p-values against CSS.
Default: 1000.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}

\item{verbose}{Whether to output message? Default: TRUE.}

\item{use_altExp}{Whether to use altExps in \code{SingleCellExperiment}.
Default: TRUE. The activate matrix stored in altExps(x). So when perform
cell-type specific analysis on regulon activate matrix, this parameter must be
TRUE. When performing expression level analysis, set this parameter to FALSE.}

\item{use_assay}{Setting the assays used for cell-type specific analysis. For
regulon analysis, it could be 'viper', 'aucell' or 'ssgsea'. For gene expression
level, it could be 'counts' or 'logcounts'. Default: the first assay.}
}
\value{
A \code{data.frame} with differential CSS for each TFRs.
}
\description{
Identifying cell-type specific transcription factor regulons for indicated cluster
}
