% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorMat.R
\name{getCorMat,matrix-method}
\alias{getCorMat,matrix-method}
\title{Calculating the weighted TF-Target values}
\usage{
\S4method{getCorMat}{matrix}(
  x,
  use_assay = 1L,
  method = c("spearman", "genie3", "pcor-spearman", "sclink", "pearson", "kendall"),
  use_regulator = TRUE,
  melt_cor_mat = FALSE,
  ncores = 1,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{x}{An scATGR object.}

\item{use_assay}{The i-th assays used for calculating the co-expression network.
It should be an gene expression matrix with raw counts or normalized values
such as TPM or FPKM. Using \code{\link{assays()}} to find assays in your
\code{\link{scATGR}} object. Default: 1.}

\item{method}{Methods used for weighted matrix calculation. Current version only
support four types of methods. The "spearman", "pearson" and "kendall" were common
used method in \code{\link{cor()}} function. The "genie3" for \code{\link{GENIE3}}
method from package \code{\link{GENIE3}}. The "pcor-spearman" for \code{\link{ppcor}}.
Default: "spearman".}

\item{use_regulator}{Whether to use regulators for enrichment analysis?}

\item{melt_cor_mat}{Whether to melt cor matrix? Default:TRUE}

\item{ncores}{Number of cores used for weight cor matrix calculation.}

\item{verbose}{Whether to show messages?}

\item{...}{Others parameters passed to \code{\link{GENIE3}} and \code{\link{cor}}.}
}
\description{
Calculating the weighted TF-Target values for \code{\link{ATFinder}} object
}
\examples{


}
