% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHomologGene}
\alias{getHomologGene}
\title{Get homologene from local database}
\usage{
getHomologGene(
  input_gene = NULL,
  input_tax = "human",
  output_tax = "mouse",
  output_type = c("gene_symbol", "gene_id", "both"),
  db = scATFR::homologene_data
)
}
\arguments{
\item{input_gene}{Input genes, which can be gene symbols or entrez gene ids.}

\item{input_tax}{Input gene taxonomy id or names. To browse the available organisms,
please use \code{\link{browserTaxonomy()}}.}

\item{output_tax}{Output gene taxonomy id or names. To browse the available organisms,
please use \code{\link{browserTaxonomy()}}}

\item{output_type}{Output types of genes. Default: gene_symbol.}

\item{db}{Database used in our homologene transformation.}
}
\value{
A data.frame include homologene gene information
}
\description{
Get homologene from local database
}
