% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHomologGRN}
\alias{getHomologGRN}
\title{Get gene regulatory network homologous}
\usage{
getHomologGRN(
  df,
  input_tax = "human",
  output_tax = "mouse",
  output_type = c("gene_symbol", "gene_id"),
  db = scATFR::homologene_data
)
}
\arguments{
\item{df}{The input gene regulatory network used for homologous transformation.
please use \code{\link{browserTaxonomy()}}.}

\item{output_tax}{Output gene taxonomy id or names. To browse the available organisms,
please use \code{\link{browserTaxonomy()}}.}

\item{output_type}{Output types of genes. Default: gene_symbol.}

\item{db}{Database used in our homologene transformation.}
}
\value{
A gene regulatory network of homologous genes.
}
\description{
Get gene regulatory network homologous
}
\examples{
hs_pantissue <- readRDS(system.file("./extdata/hs_pantissue_network.rds",package = "scATFR"))
mm_pantissue <- getHomologGRN(df = hs_pantissue, output_tax = "mouse")
}
