% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{getRadviz}
\alias{getRadviz}
\title{getRadviz object}
\usage{
getRadviz(
  input_frm,
  anchors = colnames(input_frm),
  trim_ratio = 0.005,
  optim_anchor = TRUE,
  recenter = NULL,
  text_color = "orangered4",
  text_size = 8,
  ...
)
}
\arguments{
\item{input_frm}{Matrix used to generate the Radviz object.}

\item{anchors}{Dimensions used to generate the map.}

\item{trim_ratio}{Whether to trim the top and bottom of each anchor which correspond to outliers.}

\item{optim_anchor}{Whether to optimize the anchor order?}

\item{recenter}{User defined start anchor.}

\item{...}{Other parameters passed to \code{\link{do.radviz}}.}

\item{label_col}{Label color, inherited from \code{\link{do.radviz}}.}

\item{label_size}{Label size, inherited from \code{\link{do.radviz}}.}
}
\value{

}
\description{
getRadviz object
}
