% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulons.R
\name{getRegulons,matrix-method}
\alias{getRegulons,matrix-method}
\alias{getRegulons,data.frame-method}
\title{Getting Transcription Factors Regulons using motif cisTarget information}
\usage{
\S4method{getRegulons}{matrix}(
  x,
  motif_ranking,
  motif_tf_anno,
  method = c("cisTarget", "GSEA"),
  use_regulator = NULL,
  ncores = 1,
  ...
)

\S4method{getRegulons}{data.frame}(
  x,
  motif_ranking,
  motif_tf_anno,
  method = c("cisTarget", "GSEA"),
  use_regulator = NULL,
  ncores = 1,
  ...
)
}
\arguments{
\item{x}{A matrix or data.frame containing gene regulatory network.}

\item{motif_ranking}{A \code{\link{rankingRcisTarget}} object or a \code{\link{matrix}} or
\code{\link{data.frame}} containing motif ranking matrix, in which rows
corresponding to motif ids and columns corresponding to genes.}

\item{motif_tf_anno}{A \code{\link{data.table}} with motif binding TF information.}

\item{method}{Methods used for TFR identification.}

\item{...}{Others parameter passed to \code{\link{grnToLists}}, \code{\link{grnRcisTarget}}
\code{\link{doMotifRankGSEA}}}
}
\value{
A list of TF regulons.
}
\description{
Getting Transcription Factors Regulons using motif cisTarget information
}
