% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulons.R
\name{grnRcisTarget}
\alias{grnRcisTarget}
\title{Filtering Gene Regulatory Networks using motif information}
\usage{
grnRcisTarget(
  gene_list,
  motif_rank,
  motif_tf_anno,
  min_target_num = 5,
  max_target_num = 500,
  ncores = 1,
  ...
)
}
\arguments{
\item{gene_list}{A list of TF regulons used for filtering.}

\item{motif_rank}{motif ranking object import from RcisTarget.}

\item{motif_tf_anno}{A data.table including motif and TF annotation.}

\item{min_target_num}{Minimum number of targets in your final regulons.}

\item{max_target_num}{Maximum number of targets in your final regulons.}

\item{ncores}{Number of cores used for processing.}

\item{...}{Others parameters passed to \code{\link{cisTarget}}}
}
\value{
A list of TF regulons
}
\description{
Filtering Gene Regulatory Networks using motif information
}
