% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulons.R
\name{grnToLists}
\alias{grnToLists}
\title{Transform Gene Regulatory Network to regulons}
\usage{
grnToLists(
  grn_mat,
  use_regulator = NULL,
  cutoff = 0.05,
  use_target = NULL,
  rank_type = c("pos", "neg")
)
}
\arguments{
\item{grn_mat}{A matrix included weighted Gene Regulatory Network (GRN). Rows
were regulators (eg. TFs) and columns were target genes.}

\item{use_regulator}{The regulator used for regulons.}

\item{cutoff}{Percentage of top ranked genes used for constructing regulons. Default: 0.05.}

\item{use_target}{Keep wanted target genes in regulons.}

\item{rank_type}{Get positive enriched or negative enriched targets? Default: pos.}
}
\value{
A list of TF regulons.
}
\description{
Transform Gene Regulatory Network to regulons
}
