% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAtfrObject.R
\name{import10XGenomics}
\alias{import10XGenomics}
\title{Importing from 10X genomics data}
\usage{
import10XGenomics(path)
}
\arguments{
\item{path}{The directory of 10X genomics output.}
}
\value{
An object.
}
\description{
Importing from 10X genomics data
}
