% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGRNs.R
\name{inferGRNs,SingleCellExperiment-method}
\alias{inferGRNs,SingleCellExperiment-method}
\alias{inferGRNs,Seurat-method}
\alias{inferGRNs,ExpressionSet-method}
\title{Inferring Gene Regulatory Network (GRN) from Single-cell Expression Data}
\usage{
\S4method{inferGRNs}{SingleCellExperiment}(
  x,
  regulon_list = NULL,
  use_assay = 1L,
  method = c("pcor", "genie3", "sclink", "sincerities"),
  time_col = NULL,
  ncores = 1,
  ...
)

\S4method{inferGRNs}{Seurat}(
  x,
  use_assay = 1L,
  method = c("sctenifoldnet", "genie3", "sclink", "sincerities", "pcor", "spearman"),
  use_regulator = NULL,
  time_col = NULL,
  ...
)

\S4method{inferGRNs}{ExpressionSet}(
  x,
  method = c("sctenifoldnet", "genie3", "sclink", "sincerities", "pcor", "spearman"),
  use_regulator = NULL,
  time_col = NULL,
  ...
)
}
\arguments{
\item{x}{An object used for inferring GRNs. We support object with the
following types: \code{\link{matrix}}, \code{\link{SingleCellExperiment}},
\code{\link{Seurat}}, \code{\link{monocle3}}.}

\item{use_assay}{Assay used for GRN inferring, can be number or assay names.
default: the first assay.}

\item{method}{Methods used for weighted matrix calculation. Current version only
support four types of methods. The "genie3" for \code{\link{GENIE3}} method from package \code{\link{GENIE3}}.
The "sclink" for package \code{\link{sclink}}. The "pcor" for \code{\link{ppcor}} package. The "spearman",
were common used method in \code{\link{cor()}} function.}

\item{time_col}{A numeric \code{vector} or a one column \code{data.frame} contains time points for each sample.
If the input object x is a \code{\link{SingleCellExperiment}} or \code{\link{Seurat}} object with `colData`
information, the `time_col` could be the name of column with time information in `colData(x)`. The time points
must be numeric (eg. 0, 12, 24, 48) and t least 3 time points needed. Only useful for `sincerities` method.}

\item{ncores}{Number of cores used for parallel calculation. Using \code{\link{parallel::detectCores()}}
to detect the cores on your computer.Default: 1.}

\item{...}{Other parameters passed to \code{\link{inferWeightExpMat}}}

\item{use_regulator}{Regulators used for weighted matrix calculation, usually TFs. Default: NULL, stands for all
genes were used.}
}
\description{
Inferring Gene Regulatory Network (GRN) from Single-cell Expression Data
}
