% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonPathwaySimilarity.R
\name{inferRegulonFun,SingleCellExperiment-method}
\alias{inferRegulonFun,SingleCellExperiment-method}
\title{Inferring the regulon function using pathway information}
\usage{
\S4method{inferRegulonFun}{SingleCellExperiment}(
  x,
  pathway_list,
  regulon_list = NULL,
  return_items = c("all", "similarity"),
  method = c("asymptotic", "bootstrap", "mca", "exact"),
  nperm = 1000,
  ncores = 1,
  sig_fdr = 0.05
)
}
\arguments{
\item{x}{Object used for function inferring.}

\item{pathway_list}{A list of pathway gene sets. Names of the list were pathway names,
while the value of each pathway is genes.}

\item{method}{Method used compute p-values in \pkg{jaccard} package. The "asymptotic"
is the fastest, while the "bootstrap" is better but a little bit slow. The "exact"
solution is slow for a moderately large vector. Default: "asymptotic".}

\item{nperm}{Number of . Used only for "bootstrap" method.}

\item{ncores}{Number of cores used for parallel calculation.}

\item{sig_fdr}{Set the cutoff of similarity significance between \code{regulon_list} and \code{pathway_list}.}
}
\value{
a list of similarity test \code{data.frame} or a \code{matrix} of
normalized similarity scores.
}
\description{
This method estimate the function similarities of TFRs by using
Jaccard/Tanimoto similarity coefficients.
}
