% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGRNs.R
\name{inferWeightExpMat}
\alias{inferWeightExpMat}
\title{Inferring weighted expression matrix}
\usage{
inferWeightExpMat(
  exp_mat,
  method = c("puic", "genie3", "pcor", "sincerities"),
  use_regulator = NULL,
  exp_cutoff = 0,
  col_time = NULL,
  ncores = 1,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{exp_mat}{Matrix. Numeric matrix of gene expression data with rows were gene names and columns were samples.}

\item{method}{Methods used for weighted matrix calculation. Current version only
support four types of methods. The "genie3" for \code{\link{GENIE3}} method from package \code{\link{GENIE3}}.
The "pcor" for \code{\link{ppcor}} package. The "spearman", were common used method in \code{\link{cor()}} function.}

\item{use_regulator}{Regulators used for weighted matrix calculation, usually TFs. Default: NULL, stands for all
genes were used.}

\item{exp_cutoff}{Numeric. The number or percentage of cells expressed for each gene. Default: 1, at least one cell.}

\item{col_time}{A two columns data.frame. The first column must be sample names which should be identical
with sample names in \code{exp_mat}. The second column must be time points for each sample which should
be numeric (eg. 0, 12, 24, 48). If you do not have exactly time information for your cell types, you can
just set time points from zero (eg. 0, 1, 2, 3, 4, ...). Note: at least 3 time points needed.}

\item{ncores}{Number of cores used for parallel calculation. Using \code{\link{parallel::detectCores()}}
to detect the cores on your computer.Default: 1.}

\item{verbose}{Whether to show messages?}

\item{...}{Other parameters passed to \pkg{GENIE3} \pkg{sincerities} and \pkg{sclink}.}
}
\value{
A matrix with weighted regulator-target correlation result.
}
\description{
Inferring weighted expression matrix
}
