% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerGRNs.R
\name{integerGRNs,matrix-method}
\alias{integerGRNs,matrix-method}
\alias{integerGRNs,SingleCellExperiment-method}
\title{Integer the gene regulatory network with pre-defined regulators and
targets regulatory network}
\usage{
\S4method{integerGRNs}{matrix}(
  x,
  baseNets = NULL,
  cutoff = NULL,
  weight = c(1, 1),
  norm_weight = FALSE,
  keep_neg_regu = FALSE,
  keepWeight = TRUE,
  beta0 = 0,
  ncores = 1,
  verbose = interactive(),
  ...
)

\S4method{integerGRNs}{SingleCellExperiment}(
  x,
  baseNets = NULL,
  use_assay = NULL,
  cutoff = NULL,
  weight = c(1, 1),
  norm_weight = FALSE,
  keep_neg_regu = TRUE,
  beta0 = 0,
  ncores = 1,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{x}{A gene regulatory \code{matrix} or \code{SingleCellExperiment} object.
For a matrix, rows represent for regulators (eg, TFs) and columns represent
for targets.}

\item{baseNets}{A data.frame including the background regulatory network used
for integration. There are three columns in this data.frame: tf, target and weight.
Default: NULL. No background regulatory network. Default: NULL.}

\item{cutoff}{The cutoff used for filtering the targets for each TF regulon.}

\item{weight}{The weight of each data.frame.Default: 1. Equal for each data.frame.}

\item{norm_weight}{Whether to normalize the weight for each data.frame? Default:FALSE.}

\item{keep_neg_regu}{Whether to keep negative weight for output? Default:FALSE}

\item{keepWeight}{Whether to keep merged weight in the final output. Default: TRUE.}

\item{beta0}{Numeric. The sparsity parameter.}

\item{ncores}{Numeric. The number of cores used.}

\item{verbose}{Whether to print messages on screen?}

\item{...}{Other parameters passed to \code{\link{integraNets}}.}

\item{use_assay}{The GRN assays used for integration.}
}
\value{
A list or an object with regulon list.
}
\description{
Integer the gene regulatory network with pre-defined regulators and
targets regulatory network
}
