% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapHomologList}
\alias{mapHomologList}
\title{Title}
\usage{
mapHomologList(
  input_gene,
  input_tax = "human",
  output_tax = "mouse",
  isListNameGene = FALSE,
  ...
)
}
\arguments{
\item{input_gene}{Input gene list}

\item{input_tax}{Input gene taxonomy id or names. To browse the available organisms,
please use \code{\link{browserTaxonomy()}}}

\item{output_tax}{Output gene taxonomy id or names. To browse the available organisms,
please use \code{\link{browserTaxonomy()}}}

\item{isListNameGene}{Whether the gene names is gene symbols?}
}
\value{
A gene list with homolog transformed list.
}
\description{
Title
}
