% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotRegulonReducedDim}
\alias{plotRegulonReducedDim}
\title{Plot reduced dimensions}
\usage{
plotRegulonReducedDim(object, alt_assay = 1L, dimred, ...)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{alt_assay}{String or integer scalar specifying which assay in
\code{altExp(x)} containing the input data.}

\item{dimred}{A string or integer scalar indicating the reduced dimension
result in \code{\link{reducedDims(object)}} to plot.}

\item{...}{Other parameters passed to \code{\link{plotReducedDim}}.}
}
\value{

}
\description{
Plot regulon level activation reduced dimension results stored in a
\code{\link{SingleCellExperiment}} object.
}
