% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{RadvizEnrichPlot,data.frame-method}
\alias{RadvizEnrichPlot,data.frame-method}
\title{RadvizEnrichPlot}
\usage{
\S4method{RadvizEnrichPlot}{data.frame}(
  x,
  anchors = colnames(x),
  color_by = NULL,
  return_rv = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or objects used for \code{\link{Radviz}} plot.}

\item{anchors}{Anchors used to plot.}

\item{color_by}{Groups used to label the color.}

\item{return_rv}{Only return \code{\link{getRadviz}} object, not generate figures.}

\item{...}{Other parameters passed to \code{\link{getRadviz}}.}
}
\description{
RadvizEnrichPlot
}
