% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonActivity.R
\name{regulonActivity,SingleCellExperiment-method}
\alias{regulonActivity,SingleCellExperiment-method}
\alias{regulonActivity,matrix-method}
\title{Performing regulon activity evaluation}
\usage{
\S4method{regulonActivity}{SingleCellExperiment}(
  x,
  gene_list = NULL,
  method = c("viper", "aucell", "ssgsea"),
  normalization = TRUE,
  min_target_num = 5L,
  max_target_num = 500L,
  use_assay = 1L,
  use_regulon = 1L,
  norm_act = FALSE,
  norm_coe = 1,
  ncores = 1,
  verbose = FALSE,
  ...
)

\S4method{regulonActivity}{matrix}(
  x,
  gene_list,
  method = c("viper", "ssgsea", "aucell"),
  normalization = TRUE,
  min_target_num = 5,
  max_target_num = 500,
  norm_act = FALSE,
  scale_coe = 10,
  ncores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An object with gene expression data used for regulon activity evaluation,
which can be \code{matrix}, \code{SingleCellExperiment}, \code{Seurat} and \code{ExpressionSet}.}

\item{gene_list}{List. Regulons list used for activity evaluation, names of each regulons
should be Transcription Factors(TFs).}

\item{method}{Method used to calculate the active score of regulons for each sample.}

\item{normalization}{Logical, whether to normalize the activate score?}

\item{min_target_num}{Minimum number of genes in your regulons.}

\item{max_target_num}{Maximum number of genes in your regulons.}

\item{use_assay}{For object with \code{SummarizedExperiment} class, the use_assay
argument can be used to select the assay containing the expression data we want as input.}

\item{use_regulon}{The regulon assay used for activity estimation. Default: 1, the first one.}

\item{norm_act}{Whether to scale the activity score? Default: FALSE.}

\item{norm_coe}{The coefficient used for activity score normalization.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}

\item{verbose}{Gives information about each calculation step. Default: TRUE.}

\item{...}{Other parameters passed to \code{\link{runViper}}, \code{\link{runAUCell}} and
\code{\link{runGSVA}}.}
}
\value{

}
\description{
Performing regulon activity evaluation
}
