% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction.R
\name{regulonTSNE,matrix-method}
\alias{regulonTSNE,matrix-method}
\alias{regulonTSNE,SingleCellExperiment-method}
\title{Perform t-SNE on regulon activation data}
\usage{
\S4method{regulonTSNE}{matrix}(
  x,
  use_regulon = NULL,
  ncomponents = 2,
  ntop = 500,
  scale = FALSE,
  transposed = FALSE,
  perplexity = NULL,
  normalize = TRUE,
  theta = 0.5,
  external_neighbors = FALSE,
  use_fitsne = FALSE,
  ...
)

\S4method{regulonTSNE}{SingleCellExperiment}(
  x,
  alt_assay = 1L,
  use_regulon = NULL,
  ncomponents = 2,
  ntop = 500,
  scale = FALSE,
  perplexity = NULL,
  normalize = TRUE,
  theta = 0.5,
  external_neighbors = FALSE,
  use_fitsne = FALSE,
  dimred = NULL,
  n_dimred = NULL,
  name = "TSNE",
  ...
)
}
\arguments{
\item{x}{A numeric \code{matrix} where rows are TF regulons and columns were
cells, Alternatively, a \code{\link{SingleCellExperiment}} object containing
such matrix in \code{altExp(x)} is also supported.}

\item{use_regulon}{Character vectors indicating the regulons used for TSNE
analysis. Same as \code{subset_row} in \code{\link{calculateTSNE()}}.}

\item{ncomponents}{Numeric scalar indicating the number of principal
components to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest
variances to use for dimensional reduction.}

\item{scale}{Logical scalar, should the expression values be standardized?}

\item{transposed}{Logical scalar, is x transposed with cells in rows?}

\item{perplexity}{Numeric scalar defining the perplexity parameter,
see \code{\link{?Rtsne}} for more details.}

\item{normalize}{Logical scalar indicating if input values should be scaled
for numerical precision, see \code{\link{normalize_input}}.}

\item{theta}{Numeric scalar specifying the approximation accuracy of the
Barnes-Hut algorithm, see \code{\link{Rtsne}} for details.}

\item{external_neighbors}{Logical scalar indicating whether a nearest
neighbors search should be computed externally with \code{\link{findKNN}}.}

\item{use_fitsne}{Logical scalar indicating whether \code{\link{fitsne}}
should be used to perform t-SNE.}

\item{...}{Other parameters passed to \code{\link{calculateTSNE}} and
\code{\link{runTSNE}}.}

\item{alt_assay}{String or integer scalar specifying which assay in
\code{altExp(x)} containing the input data.}

\item{dimred}{String or integer scalar specifying the existing dimensional
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{name}{String specifying the name to be used to store the result in
the \code{\link{reducedDims()}} of the output.}
}
\description{
Perform t-stochastic neighbour embedding (t-SNE) on regulons by cells
activation matrix
}
