% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reLevel,SingleCellExperiment-method}
\alias{reLevel,SingleCellExperiment-method}
\title{Reset the levels of column data from object}
\usage{
\S4method{reLevel}{SingleCellExperiment}(x, colable, level_order)
}
\arguments{
\item{x}{The SingleCellExperiment object.}

\item{colable}{The column in colData(x) used for re-order.}

\item{level_order}{The new orders.}
}
\value{
An object
}
\description{
Reset the levels of column data from object
}
