% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGRNs.R
\name{run_scTenifoldNet}
\alias{run_scTenifoldNet}
\title{Run scTenifoldNet to infer Gene Regulatory network}
\usage{
run_scTenifoldNet(
  exp_mat,
  use_regulator = NULL,
  nCells = 500,
  nNet = 100,
  nComp = 10,
  ...
)
}
\arguments{
\item{exp_mat}{A gene by sample expression matrix.}

\item{use_regulator}{Regulators used for weighted matrix calculation, usually TFs. Default: NULL, stands for all
genes were used.}

\item{nCells}{An integer value. The number of cells to subsample each time to generate a network. passed to
\code{\link{makeNetworks}}.}

\item{nNet}{An integer value. The number of networks based on principal components regression to generate.
passed to \code{\link{makeNetworks}}.}

\item{nComp}{An integer value. The number of principal components in PCA to generate the networks.
Should be greater than 2 and lower than the total number of genes. passed to \code{\link{makeNetworks}}.}

\item{...}{Other parameters passed to \code{\link{makeNetworks}} \code{\link{tensorDecomposition}}}
}
\value{
A weighted matrix for gene regulation relationship.
}
\description{
Run scTenifoldNet to infer Gene Regulatory network
}
