% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGRNs.R
\name{run_sincerities}
\alias{run_sincerities}
\title{Calculating GRNs with SINCERITIES}
\usage{
run_sincerities(exp_mat, col_data, ...)
}
\arguments{
\item{exp_mat}{A gene by sample expression matrix.}

\item{col_data}{A two columns data.frame. The first column must be sample names which should be identical
with sample names in \code{exp_mat}. The second column must be time points for each sample which should
be numeric (eg. 0, 12, 24, 48). If you do not have exactly time information for your cell types, you can
just set time points from zero (eg. 0, 1, 2, 3, 4, ...). Note: at least 3 time points needed.}

\item{...}{Other parameters passed to \code{SINCERITIES}, \code{SINCERITIES_PLUS} and
\code{final_ranked_predictions}.}
}
\value{
A list containing the following information:
\itemize{
\item adj_matrix: \emph{m} by \emph{m} matrix containing the weights of
regulatory edges. The larger \code{adj_matrix[i,j]} indicates higher confidence
that the corresponding edge exists (i.e., gene \emph{i} regulating gene \emph{j}).
\item final_table: a \code{data.frame} containing the regulation relationships
between source genes and targets and its regulatory directions.}
}
\description{
Calculating GRNs with SINCERITIES
}
\author{
\code{SINCERITIES} were first created by Nan Papili Gao and R version implemented
by Ziyi Hua. Institute for Chemical and Bioengineering.ETH Zurich. E-mail: nanp@ethz.ch.\cr
Puwen Tan Integrated the main functions from \code{SINCERITIES} into \code{run_sincerities()}, which
was included in \pkg{scATFR} package.
}
