% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runATFR.R
\name{runATFR}
\alias{runATFR}
\title{Runing ATFR in one function}
\usage{
runATFR(
  expression_data,
  col_data = NULL,
  run_steps = c(0, 1, 2, 3, 4),
  organism = c("human", "mouse"),
  use_regulon = c("dorothea", "cisTarget"),
  grn_method = c("genie3", "sincerities", "sclink", "pcor", "spearman"),
  activity_method = c("viper", "aucell", "gsva", "ssgsea", "zscore", "plage"),
  ncores = 1,
  file_out = "./",
  ...
)
}
\arguments{
\item{expression_data}{expression matrix}

\item{col_data}{sample annotation matrix}

\item{run_steps}{Run from this step}

\item{organism}{species}

\item{use_regulon}{regulon source}

\item{grn_method}{method used for regulon calculation}

\item{activity_method}{method for regulon activity}

\item{ncores}{number of cores}

\item{file_out}{path to output files.}

\item{...}{others parameters}
}
\value{

}
\description{
Runing ATFR in one function
}
