% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonActivity.R
\name{runAUCell}
\alias{runAUCell}
\title{Running AUCell with Transcription Factor Regulons}
\usage{
runAUCell(
  exp_mat,
  gene_list,
  ncores = 1,
  normalization = TRUE,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{exp_mat}{An expression matrix}

\item{gene_list}{A list of regulons.}

\item{ncores}{Number of cores used.}

\item{normalization}{}

\item{...}{Other parameters passed to \code{\link{AUCell_buildRankings}} and
\code{\link{AUCell_calcAUC}}}
}
\value{
A matrix with regulons by cell activity matrix
}
\description{
Running AUCell with Transcription Factor Regulons
}
