% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonActivity.R
\name{runGSVA}
\alias{runGSVA}
\title{Running GSVA with Transcription Factor Regulons}
\usage{
runGSVA(
  exp_mat,
  gene_list,
  method = c("gsva", "ssgsea", "zscore", "plage"),
  kcdf = c("Poisson", "Gaussian", "none"),
  min_target_num = 5,
  max_target_num = 500,
  ...
)
}
\arguments{
\item{exp_mat}{An expression matrix}

\item{gene_list}{A list of regulons.}

\item{method}{Method to employ in the estimation of gene-set enrichment scores per sample.}

\item{kcdf}{Character string denoting the kernel to use during the non-parametric estimation
of the cumulative distribution function of expression levels across samples when method="gsva".}

\item{min_target_num}{Minimum number of genes in your regulons.}

\item{max_target_num}{Maximum number of genes in your regulons.}

\item{...}{Other parameters passed to \code{\link{gsva()}}.}
}
\value{
A regulon by sample matrix with activate score.
}
\description{
Running GSVA with Transcription Factor Regulons
}
