% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction.R
\name{runReduceDim,SingleCellExperiment-method}
\alias{runReduceDim,SingleCellExperiment-method}
\alias{runReduceDim,Seurat-method}
\alias{runReduceDim,matrix-method}
\title{Performing dimension reduction analysis}
\usage{
\S4method{runReduceDim}{SingleCellExperiment}(
  x,
  alt_exp = NULL,
  exprsVal = 1L,
  method = c("PCA", "TSNE", "UMAP"),
  nPCs = 50,
  scale = FALSE,
  ...
)

\S4method{runReduceDim}{Seurat}(
  x,
  method = c("PCA", "TSNE", "UMAP"),
  nPCs = 50,
  scale = FALSE,
  ...
)

\S4method{runReduceDim}{matrix}(
  x,
  method = c("PCA", "TSNE", "UMAP"),
  nPCs = 50,
  scale = FALSE,
  ...
)
}
\arguments{
\item{alt_exp}{String or integer scalar specifying an alternative experiment containing the input data.
Can be obtained by \code{\link{altExpNames(x)}}.}

\item{exprsVal}{Integer scalar or string indicating which assay of x contains the expression values.
Can be obtained by \code{\link{assayNames(x)}}. For regulon activate matrix, use \code{\link{assayNames(altExp(x))}}}

\item{method}{String or integer scalar specifying the dimensionality reduction method to use.}

\item{nPCs}{Numeric scalar indicating the number of principal components to obtain. Same as \code{ncomponents} in
\pkg{scater}.}

\item{scale}{Logical scalar, should the expression values be standardized?}

\item{...}{Other parameters passed to \code{\link{calculatePCA()}}, \code{\link{calculateTSNE()}} and
\code{\link{calculateUMAP()}}.}
}
\value{
A \code{SingleCellExperiment} object or \code{matrix} with reduced dimension information.
}
\description{
Perform principal components analysis (PCA), t-stochastic neighbour embedding (t-SNE) and for the cells,
uniform manifold approximation and projection (UMAP) based on the data in a \code{matrix} \code{SingleCellExperiment}
or \code{Seurat} object.
}
