% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCENIC.R
\name{runSCENIC,matrix-method}
\alias{runSCENIC,matrix-method}
\alias{runSCENIC,SingleCellExperiment-method}
\title{Running simplified version of SCENIC}
\usage{
\S4method{runSCENIC}{matrix}(
  x,
  motif_ranking,
  motif_annotation,
  use_regulator = NULL,
  grn_tab = NULL,
  top_fraction = 0.05,
  nesThreshold = 0L,
  minSize = 5L,
  maxSize = 500L,
  ncores = 1,
  verbose = interactive()
)

\S4method{runSCENIC}{SingleCellExperiment}(
  x,
  motif_ranking,
  motif_annotation,
  use_assay = 1L,
  use_grn = "genie3",
  use_regulator = NULL,
  top_fraction = 0.05,
  nesThreshold = 0,
  minSize = 5L,
  maxSize = 500L,
  verbose = interactive()
)
}
\arguments{
\item{x}{The expression matrix with normalized values where rows were
genes and columns were cells. It can be a matrix or a 
\code{\link{SingleCellExperiment}} object with such expression matrix.}

\item{motif_ranking}{The motif ranking object including \code{.feather} files 
import from \pkg{RcisTarget} database using \code{\link{importRankings()}} 
method.}

\item{motif_annotation}{The motif annotation database containing the 
annotations of the motif to transcription factors.}

\item{use_regulator}{Character. The TFs used for regulon analysis.}

\item{grn_tab}{The weighted TF regulatory matrix where rows were TF genes and 
columns were target genes.}

\item{top_fraction}{The percentage of top ranked targets used for initial 
regulon analysis. Default: 0.05.}

\item{nesThreshold}{The normalized enrichment score for motif enrichment 
result. Default: 0, all enriched result.}

\item{minSize}{The minimum number of genes in one TF regulon.}

\item{maxSize}{The maximum number of genes in one TF regulon.}

\item{ncores}{Number of cores to use. Default: 1.}

\item{verbose}{Print message while runing?}
}
\value{
A matrix or object contains TF regulon activity in each single-cell.
}
\description{
This method aims to obtain the TF regulatory active matrix by running the 
RcisTarget and AUCell from SCENIC.
}
