% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonActivity.R
\name{runViper}
\alias{runViper}
\title{Running Viper with Transcription Factor Regulons}
\usage{
runViper(
  exp_mat,
  gene_list,
  min_target_num = 5,
  normalization = TRUE,
  ncores = 1,
  ...
)
}
\arguments{
\item{exp_mat}{A \pkg{ExpressionSet} object or a gene by sample matrix with gene expression
values, such as TPMs, counts or UMIs.}

\item{gene_list}{List, gene regulons used for activate analysis.}

\item{min_target_num}{Minimum number of genes in your regulons.}

\item{normalization}{Logical, whether to normalize the activate score?}

\item{ncores}{Integer, number of cores to use (only 1 in Windows-based systems)}

\item{...}{Others parameters passed to \code{viper}.}
}
\value{
A regulon by sample matrix with activate score.
}
\description{
Running Viper with Transcription Factor Regulons
}
