% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAtfrObject.R
\name{scATFR}
\alias{scATFR}
\title{Create an object for transcription factor regulon analysis}
\usage{
scATFR(exp_data, col_data = NULL, row_data = NULL, regulons = NULL)
}
\arguments{
\item{exp_data}{Expression matrix contains expression data for individual samples.
Could be non-normalized data such as counts and UMIs or normalized data such as TPM or FPKM.}

\item{col_data}{A \code{data.frame} contains sample annotation information, such as cell types,
tissues and conditions. The row names of `col_data` must be the same with the column names of
`exp_data`. Default: null.}

\item{row_data}{A \code{data.frame} contains feature annotation information. Could be gene names
or gene types. The row names of `row_data` must be the same with the row names of `exp_data`.
Default: null.}

\item{regulons}{A list of transcription factor regulons (TFRs).}
}
\value{
A \code{\link{SingleCellExperiment}} object.
}
\description{
Create an \code{\link{SingleCellExperiment}} object for transcription factor regulon (TFR) analysis
}
\examples{
library(scATFR)
#load example data
expression_data <- readRDS(system.file("extdata", "mouse_HSC_formation_expression.rds", package = "scATFR"))
colData <- readRDS(system.file("extdata", "mouse_HSC_formation_colData.rds", package = "scATFR"))
#create atfr object.
atfr <- scATFR(exp_data = expression_data, col_data=colData)


}
