% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCERITIES_GRNs.R
\name{SINCERITIES}
\alias{SINCERITIES}
\alias{SINCERITIES_PLUS}
\title{SINCERITIES}
\usage{
SINCERITIES(DATA, distance = 1, method = 1, noDIAG = 0, SIGN = 1)

SINCERITIES_PLUS(DATA, noDIAG = 0, SIGN = 1, CV_nfolds = 5)
}
\arguments{
\item{DATA}{a list containing the following information:\cr
\itemize{
\item \code{singleCELLdata}: list of length \emph{n}, where \emph{n} is the
number of capture time points. \code{DATA$singleCELLdata[[k]]} is a \emph{s_k}
by \emph{m} \code{matrix}(\code{data.frame}) containing observed expression
levels of \emph{m} genes in \emph{s_k} single cells.
\item \code{totDATA}: \emph{S} by \emph{m} matrix, where \emph{S} is the
total number of single cells (i.e., \emph{S=s_1+s_2+...+s_n} where \emph{n}
the number of capture time points) and \emph{m} is the number of genes.
\item \code{time}: vector of length \emph{n} containing the cell capture
time points or time-stamps.}}

\item{distance}{This parameter selects the distribution distance. 1 stands
for KS (Kolmogorov-Smirnov), 2 stands for CM (Cramer-von Mises), 3 stands
for AD (Anderson-Darling), 4 stands for Mean Expression Difference. Default: 1.}

\item{method}{This parameter selects the regularization regression strategy.
1 stands for RIDGE, 2 stands for ELASTIC-NET with automatic detection of
optimal alpha parameter, 3 stands for LASSO, 4 stands for ELASTIC-NET with
manual selection of alpha parameter. Default: 1.}

\item{noDIAG}{This parameter selects whether the auto-regulatory edge is
inferred. 0 stands for GRN contains no auto-regulatory edge; 1 stands for
GRN contain auto-regulatory edge. Default: 0}

\item{SIGN}{This parameter selects whether the sign / mode of the gene
regulations is inferred. 0 stands for unsigned GRN; 1 stands for signed GRN.
Default: 1.\cr
\code{SINCERITIES} uses partial correlation analysis where a positive (negative)
correlation is taken as an indication of activation (repression).}

\item{CV_nfolds}{Defines a partition of the data into \code{CV_n_folds} disjoint
subsets for the cross validation. DEFAULT: 5.}
}
\value{
A list containing the following information:
\itemize{
\item adj_matrix: \emph{m} by \emph{m} matrix containing the weights of
regulatory edges. The larger \code{adj_matrix[i,j]} indicates higher confidence
that the corresponding edge exists (i.e., gene \emph{i} regulating gene \emph{j}).
\item DISTANCE_matrix: \emph{n-1} by m matrix containing the (normalized)
distribution distance (DD) computed during the network inference.}
}
\description{
SINCERITIES is a novel computational method for inferring
gene regulatory network (GRN) from time-stamped cross-sectional
single-cell expression data. SINCERITIES_PLUS function is an extension of
the default version of SINCERITIES, designed for single cell datasets with
fewer than five time points.
}
\author{
\code{SINCERITIES} were created by Nan Papili Gao and R version implemented
by Ziyi Hua. Institute for Chemical and Bioengineering.ETH Zurich.
E-mail: nanp@ethz.ch.\cr
Puwen Tan Integrated the \code{SINCERITIES} into \pkg{scATFR}.
}
