% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{summarizeClusterValue}
\alias{summarizeClusterValue}
\title{Summarizing the values for genes in each cluster}
\usage{
summarizeClusterValue(
  object,
  clust_name,
  value = c("mean", "median", "percent"),
  gene_name = NULL,
  scale = FALSE,
  use_assay = 1L
)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{clust_name}{One column name in colData(x)(for \code{SingleCellExperiment})
clusters used for summarizing the value.}

\item{value}{The values calculated for each cell cluster. Default: mean.}

\item{gene_name}{Gene names used for summarizing. Default: NULL. All genes in
object.}

\item{scale}{Whether to scale the value?}

\item{use_assay}{Setting the assays with the expression data. It could be
'counts', 'logcounts' or 'TPM'. Default: the first assay.}
}
\value{
A matrix with average values for each cluster.
}
\description{
Summarizing the values for genes in each cluster
}
