% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{summarizeClusterValueRegulon}
\alias{summarizeClusterValueRegulon}
\title{Summarizing the values for regulons in each cluster}
\usage{
summarizeClusterValueRegulon(
  object,
  clust_name,
  value = c("mean", "median", "percent"),
  regulon_name = NULL,
  scale = FALSE,
  use_assay = 1L
)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{clust_name}{One column name in colData(x)(for \code{SingleCellExperiment})
specific the clusters used for mean value calculation.}

\item{value}{The values calculated for each cell cluster. Default: mean.}

\item{regulon_name}{Gene names used for mean value calculation. Default: NULL.
All genes in object.}

\item{scale}{Whether to scale the value?}

\item{use_assay}{Setting the alternative assays with the regulon activate
score matrix. It could be 'viper', 'aucell' or 'ssgsea'. Default: the first assay.}
}
\value{
A matrix with average values for each cluster.
}
\description{
Summarizing the values for regulons in each cluster
}
