% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tfrFromDoRoThEA}
\alias{tfrFromDoRoThEA}
\title{Get TFRs From DoRoThEA}
\usage{
tfrFromDoRoThEA(
  input,
  levels = c("A", "B", "C", "D", "E"),
  min_target_num = 5,
  max_target_num = 1000
)
}
\arguments{
\item{input}{The regulons \code{data.frame} from \pkg{dorothea} package}

\item{levels}{The levels of data used as regulons.}

\item{min_target_num}{Miniumn target gene bumber.}

\item{max_target_num}{Maxiumn target gene bumber.}
}
\description{
Get TFRs From DoRoThEA
}
