% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tfrFromRcisTarget}
\alias{tfrFromRcisTarget}
\title{extract TF regulons from Rcistarget database}
\usage{
tfrFromRcisTarget(
  motifRnking,
  motif_anno,
  cutoff = 0.01,
  use_TFs = NULL,
  motif_level = c("high", "low", "custom"),
  levels = c("directAnnotation", "inferredBy_Orthology", "inferredBy_MotifSimilarity",
    "inferredBy_MotifSimilarity_n_Orthology")
)
}
\arguments{
\item{motifRnking}{The \code{\link{Rcistarget}} motif ranking file}

\item{motif_anno}{The motif annotation file}

\item{cutoff}{The cutoff of motif enrichment.}

\item{use_TFs}{User defined TFs.}

\item{motif_level}{The levels of motif used for filtering.}

\item{levels}{The level of motif used.}
}
\value{

}
\description{
extract TF regulons from Rcistarget database
}
