% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{calcCSS}
\alias{calcCSS}
\title{Identifying cell-type specific transcription factor regulons}
\usage{
calcCSS(mat, cell_anno, pos_clust, neg_clust = NULL, nperm = 1000, ncores = 1)
}
\arguments{
\item{mat}{A \code{matrix} contains transcription factor regulons activate score.}

\item{cell_anno}{A two-column \code{data.frame} with cell name and cell cluster or cell (or stage) information.}

\item{pos_clust}{Character, cluster used to calculate the cell-type specific score (CSS).}

\item{neg_clust}{Other clusters used as controls. Default: NULL (indicating all clusters expect for 'pos_clust').}

\item{nperm}{Number of permutations used for calculating p-values against CSS. Default:1000.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}
}
\value{
A \code{data.frame} with differential CSS for each TFRs.
}
\description{
Identifying cell-type specific transcription factor regulons
}
