% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonPathwaySimilarity.R
\name{calcuModuleSimilar}
\alias{calcuModuleSimilar}
\title{Estimating the jaccard similarity between two gene lists}
\usage{
calcuModuleSimilar(
  tfr_list,
  pathway_list,
  method = c("asymptotic", "bootstrap", "mca", "exact"),
  adjust = c("BH", "bonferroni", "fdr"),
  fdr = 0.05,
  collapse_set = TRUE,
  min_overlap = 5,
  max_cluster = "auto",
  nperm = 1000,
  ncores = 1,
  verbose = interactive()
)
}
\arguments{
\item{tfr_list}{A list of transcription factor regulons. Names of the list 
were TF genes, while the value of each list is its target genes.}

\item{pathway_list}{A list of pathway gene sets. Names of the list could be 
pathway names, while the value of each pathway is genes.}

\item{method}{Method used compute p-values in \pkg{jaccard} package. The asymptotic
is the fastest, while the bootstrap is better but is a little bit slow. The exact
solution is slow for a moderately large vector. Default: asymptotic.}

\item{adjust}{Adjust methods for p-value. Default: "BH".}

\item{fdr}{The cutoff for significantly overlap terms. Default: 0.05.}

\item{collapse_set}{Collapse enriched terms with high similarities.}

\item{min_overlap}{The minimum enriched genes in pathway list. Default: 5.}

\item{max_cluster}{The maximum clusters used for merge enriched pathways. 
Default: auto, \code{sqrt(nrow(enrich_tab))}.}

\item{nperm}{Number of permutations. Used only for bootstrap method. 
Default: 1000.}

\item{ncores}{Number of cores used for parallel calculation. Default: 1.}

\item{verbose}{Whether to output information while running?}
}
\value{
a list of similarity test \code{data.frame} or a \code{matrix} of
normalized similarity scores.
}
\description{
This method estimate the similarities between two gene lists by using
Jaccard/Tanimoto similarity coefficients.
}
