% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{do_RadvizEnrichPlot}
\alias{do_RadvizEnrichPlot}
\title{do_RadvizEnrichPlot}
\usage{
do_RadvizEnrichPlot(
  rv,
  color_by = NULL,
  outline_circle = TRUE,
  plot_type = c("point", "density", "hexagonal", "buble"),
  point_size = 0.5,
  point_shape = 16,
  alpha = 1,
  ...
)
}
\arguments{
\item{rv}{A \code{\link{getRadviz}} object generate from \code{\link{do.radviz}}.}

\item{color_by}{Groups used to label the color.}

\item{outline_circle}{Whether to add circle in figure?}

\item{plot_type}{Figure types to plot.}

\item{point_size}{Label size of point in \code{\link{geom_point()}}.}

\item{point_shape}{Label shape of point in \code{\link{geom_point()}}.}

\item{...}{Other input parameters.}

\item{text_size}{Text size.}
}
\value{
a ggplot object
}
\description{
do_RadvizEnrichPlot
}
