% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{findAllCTSRegulons,SingleCellExperiment-method}
\alias{findAllCTSRegulons,SingleCellExperiment-method}
\alias{findAllCTSRegulons,matrix-method}
\title{Calculating the significant enriched regulons for each cluster}
\usage{
\S4method{findAllCTSRegulons}{SingleCellExperiment}(
  x,
  use_altExp = TRUE,
  use_assay = 1L,
  diff_by = "cluster",
  fdr = 0.05,
  css = 0.3,
  nperm = 1000,
  ncores = 1,
  ...
)

\S4method{findAllCTSRegulons}{matrix}(x, cell_anno, fdr = 0.05, css = 0.5, ...)
}
\arguments{
\item{x}{An object with gene expression data used for regulon activity evaluation,
which can be \code{matrix}, \code{SingleCellExperiment}, \code{Seurat} and \code{ExpressionSet}.}

\item{fdr}{FDR cutoff used for filtering significant regulons.}

\item{css}{CSS cutoff used for filtering significant regulons.}

\item{cell_anno}{A two-column \code{data.frame} with cell name and cell cluster or cell (or stage) information.}
}
\value{
A \code{data.frame} with differential CSS for each TFRs.
}
\description{
Calculating the significant enriched regulons for each cluster
}
