% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecific.R
\name{findCellTypeSignature,SingleCellExperiment-method}
\alias{findCellTypeSignature,SingleCellExperiment-method}
\alias{findCellTypeSignature,matrix-method}
\alias{findCellTypeSignature,ExpressionSet-method}
\alias{findCellTypeSignature,Seurat-method}
\title{Identifying cell-type specific transcription factor regulons}
\usage{
\S4method{findCellTypeSignature}{SingleCellExperiment}(
  x,
  pos_clust,
  neg_clust = NULL,
  diff_by = "cluster",
  use_altExp = TRUE,
  use_assay = 1L,
  nperm = 1000,
  ncores = 1
)

\S4method{findCellTypeSignature}{matrix}(
  x,
  cell_anno,
  pos_clust,
  neg_clust = NULL,
  nperm = 1000,
  ncores = 1
)

\S4method{findCellTypeSignature}{ExpressionSet}(
  x,
  cell_anno,
  pos_clust,
  neg_clust = NULL,
  nperm = 1000,
  ncores = 1
)

\S4method{findCellTypeSignature}{Seurat}(
  x,
  cell_anno,
  pos_clust,
  neg_clust = NULL,
  nperm = 1000,
  ncores = 1
)
}
\arguments{
\item{x}{An object with gene expression data used for regulon activity evaluation,
which can be \code{matrix}, \code{SingleCellExperiment}, \code{Seurat} and \code{ExpressionSet}.}

\item{pos_clust}{Character, cluster used to calculate the cell-type specific score (CSS).}

\item{neg_clust}{Other clusters used as controls. Default: NULL (indicating all clusters expect for 'pos_clust').}

\item{nperm}{Number of permutations used for calculating p-values against CSS. Default:1000.}

\item{ncores}{Number of threads used for parallel calculation. Default: 1.}

\item{cell_anno}{A one-column \code{data.frame}. Rows is cell name and columns is cell cluster or stages.}
}
\value{
A \code{data.frame} with differential CSS for each TFRs.
}
\description{
Identifying cell-type specific transcription factor regulons
}
