% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferGRNs.R
\name{inferGRNs,matrix-method}
\alias{inferGRNs,matrix-method}
\alias{inferGRNs,SingleCellExperiment-method}
\title{Inferring Gene Regulatory Network (GRN) from Single-cell Expression Data}
\usage{
\S4method{inferGRNs}{matrix}(
  x,
  method = c("pidc", "genie3", "pcor", "sincerities"),
  use_regulator = NULL,
  use_target = NULL,
  time_col = NULL,
  ncores = 1,
  ...
)

\S4method{inferGRNs}{SingleCellExperiment}(
  x,
  use_assay = 1L,
  method = c("pidc", "genie3", "pcor", "sincerities"),
  use_regulator = NULL,
  use_target = NULL,
  time_col = NULL,
  ncores = 1,
  ...
)
}
\arguments{
\item{x}{An object used for GRN inferring . We support object with the
following types: \code{\link{matrix}} and an object of
\code{\link{SingleCellExperiment}}.}

\item{method}{Methods used for weighted GRN inferring. Current version support
four types of methods: \code{\link{PIDC}}, \code{\link{GENIE3}},
\code{\link{PCOR}} and \code{\link{SINCERITIES}}. Default: pidc.}

\item{use_regulator}{Regulators used for weighted GRN inferring, eg. TFs.
Default: NULL, all genes in rows.}

\item{time_col}{A numeric \code{vector} or a one column \code{data.frame}
contains time points for each sample. If the input object x is a
\code{\link{SingleCellExperiment}} with `colData` information, the `time_col`
could be the column name with time information in `colData(x)`. The time points
must be numeric (eg. 0, 12, 24, 48) and at least 3 time points needed.
Only used for `sincerities` method. Default: NULL.}

\item{ncores}{Number of cores used for parallel calculation. Using
\code{\link{parallel::detectCores()}} to detect the available cores on your
computer. Default: 1.}

\item{...}{Other parameters passed to \code{\link{inferWeightExpMat}}.}

\item{use_assay}{Assay used for GRN inferring, can be \code{numeric} or
\code{character} with assay names. Default:1,  the first assay.}
}
\description{
Inferring Gene Regulatory Network (GRN) from Single-cell Expression Data
}
