% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerGRNs.R
\name{integraNets}
\alias{integraNets}
\title{Integrating weighted regulator-target regulatory networks}
\usage{
integraNets(
  grnNets,
  weight = rep(1, length(grnNets)),
  norm_weight = FALSE,
  keep_neg_regu = TRUE,
  join_type = c("full", "inner", "left", "right"),
  beta0 = 0,
  verbose = interactive()
)
}
\arguments{
\item{grnNets}{A list of weighted regulator-target regulatory networks from different
resources, for example: motif prediction, Chip-seq and Knockout experiment. For each
network in the list, there must three columns corresponding to regulator, target and
weight, respectively. If one of your network do not have weight, you can set 1 to all
the regulator-target pairs in `weight` column.}

\item{weight}{Positive number. Specific the weight of resources. default: 1}

\item{norm_weight}{Whether to normalize the weight to [0,1] of regulator-target pairs
from each resource. Default: FALSE}

\item{keep_neg_regu}{Whether to keep negative regulatory relationships between regulators
and targets. Default: TRUE.}

\item{join_type}{The merge types among different data frames. Default:full.}

\item{beta0}{Numeric. The sparsity parameter.}
}
\value{
A data.frame with integrated weighted regulator-target regulatory networks
}
\description{
Integrating weighted regulator-target regulatory networks
}
