% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{RadvizEnrichPlot,data.frame-method}
\alias{RadvizEnrichPlot,data.frame-method}
\title{RadvizEnrichPlot}
\usage{
\S4method{RadvizEnrichPlot}{data.frame}(
  x,
  anchors = colnames(x),
  color_by = NULL,
  trim_ratio = 0.005,
  norm_data = TRUE,
  optim_anchor = TRUE,
  recenter = NULL,
  text_color = "orangered4",
  text_size = 8,
  return_rv = FALSE,
  outline_circle = TRUE,
  plot_type = c("point", "density", "hexagonal", "buble"),
  point_size = 0.5,
  point_shape = 16,
  alpha = 1,
  ...
)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or objects used for \code{\link{Radviz}} plot.}

\item{anchors}{Anchors used to plot.}

\item{color_by}{Groups used to label the color.}

\item{trim_ratio}{Whether to trim the top and bottom of each anchor which correspond to outliers.}

\item{optim_anchor}{Whether to optimize the anchor order?}

\item{recenter}{User defined start anchor.}

\item{text_color}{Text color, inherited from \code{\link{getRadviz}}.}

\item{text_size}{Text size, inherited from \code{\link{getRadviz}}.}

\item{return_rv}{Only return \code{\link{getRadviz}} object, not generate figures.}

\item{outline_circle}{Whether to add circle in figure?}

\item{plot_type}{Figure types to plot. Default: point.}

\item{point_size}{Label size of points in \code{\link{geom_point()}}.}

\item{point_shape}{Label shape of points in \code{\link{geom_point()}}. Default: 16.}

\item{alpha}{The point transparency. Default: 0.8.}

\item{...}{Other parameters passed to \code{\link{getRadviz}}.}
}
\description{
RadvizEnrichPlot
}
