% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRadvizPlot.R
\name{RadvizFeaturePlot,data.frame-method}
\alias{RadvizFeaturePlot,data.frame-method}
\title{RadvizFeaturePlot}
\usage{
\S4method{RadvizFeaturePlot}{data.frame}(
  x,
  anchors = colnames(x),
  use_feature,
  return_rv = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{\link{data.frame}} or objects used for \code{\link{Radviz}} plot.}

\item{anchors}{Anchors used to plot.}

\item{use_feature}{Features used to split the screens.}

\item{return_rv}{Only return \code{\link{getRadviz}} object, not generate figures.}

\item{...}{Other parameters passed to \code{\link{getRadviz}}.}
}
\description{
RadvizFeaturePlot
}
